Two Changes to Devin
First is to not return the score based on the first opponent end state,
look for the min of any end states we can see

Second we make `make_move` do a minmax options and always take the max.
Does this make any sense? I _think_ so cause I always want to make the
move thats best for me