Bug Fixes
The first bug was cause we had the score function return None if we
weren't at the right depths. This was correct for the legacy system but
incorrect now. This led to us continuing even after I died, leading to
weird move sets without a move by us. Removing this fixes the issue!

The next was in the deepened_minmax where it was using the number of
players to adjust to depth, but we don't want to do that anymore. Instead
we can start at 2 and go down by 2 each time, since our try just has a
layer for our moves, and then a layer for all the opponent moves