Fix Alpha/Beta Bug in Devin Eval
Eval devin seems to have been alpha-betaing too much, and therefor not
checking all the right states.

I _think_ we only want to break out of the loop if we actually go past
each other, not when alpha == beta.

Also add some extra logging to all games so we get the move list Devin
decided to use

Also add a way to output a Result to an ascii tree. This was really
useful for debugging, but I think I can improve it from here