AWS Bill and Fix Panic
Penelope added some asserts to make sure the game state is consistent,
which is awesome and caught a bug I had over here!

I wasn't clearing out the moves for snakes that had died, and then would
try to apply moves that didn't make sense given the snakes that were
alive.
This played into the matching on the number of moves and the number of
players, where this got busted if a snake got eliminated. Before the
assert we likely were just having undefined behavior