Extract BoxedFactory in Axum Extractor
This is similiar to what we _wanted_ to do with Rocket a long time ago!

We have a 'middleware' that can take a Requests and give us a
`BoxedFactory` for the relevant snake. This makes it easy to define
routes that automatically get the correct snake based on the pathing and
does the right internals.

This lets me keep the `BoxedFactory` interface I've always had without
needing to manually wire up each Snake in Axum