Get tracing working in async/await code
I was holding onto a SpanEnter structy thingy over an async/await
boundry and that was confusing the whole thing!

What I needed to do was use the `instrument` method that attaches the
Span to the Future through cool magic, so that tracing understood the
correct executions even when polling on the Rust futures

We also manually set the parent of root spans we create in a new thread,
since in this case we DO want them associated with the parent thread.

We use threading to make sure we don't timeout when thinking too long,
but the new thread is conceptually part of the same unit of work.