Spread from head now uses dynamically sizes based on the board size
This works by using a vec instead of a backing array as its cell
Storage.

I need to check the benches probably but this shouldn't be a huge
difference I don't think cause this Vec is only created once and uses
the vec macro knowing the total size upfront so should be able to
reserve the whole block at once

Also hardcode hobbs to arcade maze size as a first step to getting a
version that can run multiple sizes better.