Cursed Commit
Made lots of modeling crimes.

Trying to change the shape of the tree as described in this Github
Issue: https://github.com/coreyja/battlesnake-rs/issues/29

Basically I think we need a layer for my moves, and then a layer for all
my opponent moves.

Hacked this together a tiny but, but need to remodel my Node struct into
a MyNode and OpponentsNode or something like that I think