Worked on Refactoring my generic Traits
The goal was to work on storing state in between turns. This would have
been helpful for my minimax snakes, so that they could keep the result
from the last turn around, to improve move ordering on the next turn

So far we are just adding Associated Types to some of the traits. The
goal is to add the `State` type to `BattlesnakeAI` so that we can have a
consistent State for each different Snake